fix(SLB-201): move CSS-sources and config into src
Otherwise Gatsby will pick `postcss.config.cjs` up when loading `build/styles.css`
and break due to missing plugins.
