+++ packages/ui/.storybook/preview.tsx
@@ -1,4 +1,4 @@
-import '../tailwind.css';
+import '../src/tailwind.css';
 
 import { Decorator } from '@storybook/react';
 import { IntlProvider } from '../src/utils/intl';
