+++ packages/ui/package.json
@@ -30,9 +30,9 @@
   "scripts": {
     "prep:types": "tsc --emitDeclarationOnly",
     "prep:scripts": "swc ./src -d ./build",
-    "prep:styles": "NODE_ENV=production pnpm postcss tailwind.css -o build/styles.css",
-    "prep:iframe": "NODE_ENV=production pnpm postcss iframe.css -o build/iframe.css",
-    "prep:gutenberg": "NODE_ENV=production PREFIX=gutenberg pnpm postcss tailwind.css -o build/gutenberg.css",
+    "prep:styles": "NODE_ENV=production pnpm postcss src/tailwind.css -o build/styles.css",
+    "prep:iframe": "NODE_ENV=production pnpm postcss src/iframe.css -o build/iframe.css",
+    "prep:gutenberg": "NODE_ENV=production PREFIX=gutenberg pnpm postcss src/tailwind.css -o build/gutenberg.css",
     "prep:i18n": "formatjs extract 'src/**/*.ts*' --ignore='**/*.d.ts' --ignore='**/*.stories.ts*' --out-file build/translatables.json --id-interpolation-pattern '[sha512:contenthash:base64:6]'\n",
     "build": "storybook build",
     "dev": "storybook dev -p 6006 --no-open",
