+++ packages/ui/src/tailwind.config.cjs
